rag-chatbot-backend/requirements.txt
fastapi
uvicorn
cohere
qdrant-client
beautifulsoup4
python-dotenv
openai
requests
